+++ NEWS.md
@@ -1,7 +1,9 @@
 # infer 0.1.0
 - Updated documentation with examples
 - Created `pkgdown` site materials
+- Added `obs_stat` and `shade_direction` arguments to `visualize()`
 - Adding theoretical distributions to `visualize()`
+
 
 # infer 0.0.1
 - Implemented the "intro stats" examples for randomization methods as shown on the README